requirements.txt
Flask~=3.1.0
black==24.3.0
isort==5.12.0
pytest==7.4.0
redis==5.0.1
elasticsearch==8.13.1
google-adk==1.3.0
strands-agents[litellm]==1.6.0
strands-agents-tools==0.2.5
python-dotenv==1.0.0
fastmcp>=0.1.0
